# Cache midpoints in Ptable footer for faster startup times