… src/EventStore.Core.Tests/Index/IndexV1/adding_sixteen_items_to_empty_index_map_with_four_tables_per_level_causes_double_merge.cs
… src/EventStore.Core.Tests/Index/IndexV1/corrupt_index_should.cs
… src/EventStore.Core.Tests/Index/IndexV1/index_map_should_detect_corruption.cs
… src/EventStore.Core.Tests/Index/IndexV1/opening_a_ptable_with_more_than_32bits_of_records.cs
… src/EventStore.Core.Tests/Index/IndexV1/ptable_midpoint_cache_should.cs
… src/EventStore.Core.Tests/Index/IndexV1/ptable_range_query_tests.cs
… src/EventStore.Core.Tests/Index/IndexV1/ptable_read_scenario_with_items_spanning_few_cache_segments.cs
… src/EventStore.Core.Tests/Index/IndexV1/ptable_read_scenario_with_usual_items.cs
… src/EventStore.Core.Tests/Index/IndexV1/ptable_should.cs
… src/EventStore.Core.Tests/Index/IndexV1/saving_index_with_six_items_to_a_file.cs
… src/EventStore.Core.Tests/Index/IndexV1/table_index_on_range_query.cs
… src/EventStore.Core.Tests/Index/IndexV1/table_index_on_try_get_one_value_query.cs
… src/EventStore.Core.Tests/Index/IndexV1/table_index_should.cs
… src/EventStore.Core.Tests/Index/IndexV1/table_index_with_corrupt_index_entries_should.cs
… src/EventStore.Core.Tests/Index/IndexV1/table_index_with_two_ptables_and_memtable_on_range_query.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_a_ptable_header_is_corrupt_on_disk.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_a_ptable_is_corrupt_on_disk.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_a_ptable_is_loaded_from_disk.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_creating_ptable_from_memtable.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_merging_four_ptables.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_merging_ptables.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_merging_ptables_with_entries_to_nonexisting_record.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_merging_two_ptables.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_trying_to_get_latest_entry.cs
… src/EventStore.Core.Tests/Index/IndexV1/when_trying_to_get_oldest_entry.cs
… src/EventStore.Core.Tests/Index/IndexV2/adding_four_items_to_empty_index_map_with_four_tables_per_level_causes_merge.cs
… src/EventStore.Core.Tests/Index/IndexV2/adding_four_items_to_empty_index_map_with_two_tables_per_level_causes_double_merge.cs
… src/EventStore.Core.Tests/Index/IndexV2/adding_item_to_empty_index_map.cs
… src/EventStore.Core.Tests/Index/IndexV2/adding_sixteen_items_to_empty_index_map_with_four_tables_per_level_causes_double_merge.cs
… src/EventStore.Core.Tests/Index/IndexV2/adding_two_items_to_empty_index_map_with_two_tables_per_level_causes_merge.cs