Large databases with very large index files (10+ GB) can take a lot of time to start up due to the time taken to cache the midpoints into memory. Although there are not many midpoints (up to 2^28, 24 bytes each - around 6GB max), it takes much time to load them from disk since they are not stored sequentially and the OS reads data in large blocks.

This change caches the midpoints at the end of the PTable file for faster loading times. **It will not improve speeds for existing large databases, until the next large PTable merges or if the indexes are rebuilt**

**Add a new version of PTable, V4:**
1) keep a 128 byte footer before the md5 hash
2) first 4 bytes of the footer stores an unsigned int: the number of midpoints cached in the PTable
3) Midpoints will be cached after index entries
4) Each midpoint index entry has the same size as an index entry

Midpoints will be cached during the next PTable merge (only if no index entries have been removed during the merge, due to missing database entries)
On startup, if index verification is disabled (`--skip-index-verify`) and cached midpoints are present in the file, they will be loaded directly into the midpoints memory cache.
Otherwise, they will be loaded normally.