@gregoryyoung 2017-11-15
might it make sense to mmap the cached midpoints?