@shaan1337 2017-11-15
in the worst case, it looks like the midpoints cache would be around 6GB, taking around 12s for a read speed of 500MB/s but we need to do some tests to confirm this. it could help save some memory though.