@shaan1337 2017-11-20
What are your views on handling corrupted index entries (might happen if index verification has been disabled)?

@hayley-jean and I proposed the following:
We could add some assertions like the following (if not already present):
* When reading a stream, all index entries for that stream should have same stream hash/events should be ordered
* File reading positions should be a multiple of the entry sizes

If any of these assertions fail, add a flag that forces index & chunk file verification next time Eventstore is started so that indexes can be rebuilt if corrupted.

What do you think?