@gregoryyoung 2017-11-20
When reading a stream, all index entries for that stream should have same stream hash/events should be ordered
File reading positions should be a multiple of the entry sizes

I believe both of these are already there.