@gregoryyoung 2017-11-20
https://github.com/EventStore/EventStore/blob/release-v4.0.4/src/EventStore.Core/Index/PTable.cs#L440 as example (there are bunches of similar checks)