@shaan1337 2017-11-22
# Performance tests

Tests with a 10GB PTable
**With index verification:** 78s
**Without index verification, with cached midpoints:** 0.8s
**Without index verification and without loading cached midpoints:** 62s (based on previous tests)

## Generate a 10GB PTable
```
Header: 128 bytes
Index Entries: 450M * 24 bytes = 10800000000 bytes
Number of midpoints: (2*2**(21-1))*4096 < 10800000000 = Depth: 21, 2^21 = 2097152 midpoints * 24 bytes = 50331648
PTable footer: 128 bytes
MD5 hash: 16 bytes

#Write Header
echo -n -e \\x01\\x04 >> 10gb.dat
dd if=/dev/zero bs=126 count=1 >> 10gb.dat

#Write index entries: 10299*1024*1024 + 715776 remainder bytes
dd if=/dev/urandom bs=1M count=10299 >> 10gb.dat
dd if=/dev/urandom bs=715776 count=1 >> 10gb.dat

#Midpoint entries: 50331648 bytes
dd if=/dev/urandom bs=1M count=48 >> 10gb.dat

#Footer
echo -n -e \\x01\\x04 >> 10gb.dat
#Midpoint count, 2097152
echo -n -e \\x00\\x00\\x20\\x00 >> 10gb.dat
dd if=/dev/zero bs=122 count=1 >> 10gb.dat

#MD5 Hash
dd if=/dev/urandom bs=16 count=1 >> 10gb.dat

#Check file size:
du -b 10gb.dat
10850331920     10gb.dat

#Replace a ptable in index folder with 10gb.dat (rename 10gb.dat to the index file name, in this example: 9981ceda-5c11-41f6-b53c-512463b5c0d0)
```

## Running EventStore
```
#Clear disk cache
echo 3 > /proc/sys/vm/drop_caches

#Run EventStore without index verification: 0.8s
mono ./bin/clusternode/EventStore.ClusterNode.exe --db ../db --log ../logs --skip-db-verify --skip-index-verify

[19664,15,06:43:57.964] Loading of PTable '9981ceda-5c11-41f6-b53c-512463b5c0d0' started...
[19664,15,06:43:57.968] Disabling Verification of PTable
[19664,15,06:43:57.968] Loading 2097152 cached midpoints from PTable
[19664,07,06:43:58.674] CACHED TFChunk #7-7 (chunk-000007.000000) in 00:00:01.0601896.
[19664,15,06:43:58.779] Loading PTable (Version: 4) '9981ceda-5c11-41f6-b53c-512463b5c0d0' (450000000 entries, cache depth 21) done in 00:00:00.8156375.

#Clear disk cache
echo 3 > /proc/sys/vm/drop_caches

#Run EventStore with index verification: 78s
mono ./bin/clusternode/EventStore.ClusterNode.exe --db ../db --log ../logs --skip-db-verify

[19769,16,06:44:55.132] Loading and Verification of PTable '9981ceda-5c11-41f6-b53c-512463b5c0d0' started...
[19769,12,06:44:55.460] CACHED TFChunk #7-7 (chunk-000007.000000) in 00:00:00.9763872.
[19769,16,06:46:13.113] ReadIndex is corrupted...
```